requirements.txt
streamlit
pandas
matplotlib
seaborn
python-dotenv
google-generativeai